refactor: 승자를 결정하여 출력하는 기능 구현
