+++ docs/README.md
@@ -59,10 +59,10 @@
     - [x] 폰에 대해 같은 File에 존재하는 수를 계산한고 점수 계산하기
     - [x] 각 피스별 점수를 합하기
 
-- [ ] 승패 결정
+- [x] 승패 결정
     - [x] 게임이 끝났는지 확인하기
         - [x] 왕이 죽었는지 확인하기
-    - [ ] 게임이 끝났으면 승자를 결정하기
+    - [x] 게임이 끝났으면 승자를 결정하기
 
 - [x] 데이터 베이스와 연동하기
     - [x] 데이터 베이스와 연결하기
